build: updated project commands
