+++ deploy/package.json
@@ -10,7 +10,9 @@
     "execa": "^9.5.2"
   },
   "scripts": {
-    "build": "esbuild ../contract/axelar-gmp.contract.js --platform=node --format=esm --bundle --outfile=dist/axelar-gmp.contract.bundle.js && agoric run src/init-axelar-gmp.js",
+    "build:contract": "esbuild ../contract/axelar-gmp.contract.js --platform=node --format=esm --bundle --outfile=dist/axelar-gmp.contract.bundle.js && agoric run src/init-axelar-gmp.js",
+    "build:token": "agoric run ./asset-builder/register-interchain-bank-assets.builder.js --assets='[{\"denom\":\"ibc/2CC0B1B7A981ACC74854717F221008484603BB8360E81B262411B0D830EDE9B0\",\"issuerName\":\"AXL\",\"decimalPlaces\":6}]'",
+    "build": "yarn build:contract && yarn build:token",
     "test": "yarn build && ava",
     "deploy": "yarn build && runInsideContainer=true planFile=startAxelarGmp-plan.json node deploy.js",
     "register:token": "runInsideContainer=true planFile=eval-register-interchain-bank-assets-plan.json node deploy.js"
