+++ integration/package.json
@@ -14,7 +14,7 @@
     "multicall": "node scripts/multicall.mjs"
   },
   "devDependencies": {
-    "@agoric/client-utils": "^0.2.0-u19.2",
+    "@agoric/client-utils": "0.2.0-u19.2",
     "@agoric/internal": "^0.4.0-u19.2",
     "execa": "^9.5.2",
     "ses": "^1.12.0"
